misc/hotdog-or-not/challenge/requirements.txt
Flask==3.1.1
Pillow==11.1.0
filetype==1.2.0

tensorflow==2.18.0
keras==3.9.2
numpy
h5py==3.13.0

Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.2
itsdangerous==2.2.0
click==8.1.8
blinker==1.9.0

tifffile
imagecodecs
scipy
